Create RainStage component
